# src/backend/doc/swagger.yaml
swagger: "2.0"
info:
  description: "TestBuDDy is a test management, requirement management, automation friendly tool."
  version: "1.0.0"
  title: "TestBuDDy"
  contact:
    email: "barbora.blozonova@gmail.com"
  license:
    name: "MIT licence"
host: "127.0.0.1:5000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /init-requirements:
    post:
      tags:
      - "Demo Initialization"
      summary: "Initialize test tables for requirement management"
      description: "(tables: tag, requirement)."
      produces:
      - "text/html"
      responses:
        200:
          description: "Operation successful. Requirement test tables initialized."
        500:
          description: "Operation failed."
...
  /projects/{proj_id}/requirements/{req_id}:
    get:
      tags:
      - "Requirement management"
      summary: "Get requirement for project"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "req_id"
        description: "requirement id"
        in: "path"
        type: "integer"
        required: true
      responses:
        404:
          description: "Project or requirement not found."
        200:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                $ref: "#/definitions/Requirement_full"
          description: "Operation successful. Returns requirement in JSON response."
    put:
      tags:
      - "Requirement management"
      summary: "Update requirement info"
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "req_id"
        description: "requirement id"
        in: "path"
        type: "integer"
        required: true
      - name: "Requirement info"
        in: "body"
        description: "Requirement information"
        required: true
        schema:
          type: "object"
          properties:
            content:
              type: "string"
      responses:
        200:
          description: "Operation successful. Returns updated requirement in JSON response."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                $ref: "#/definitions/Requirement_full"
        400:
          description: "Invalid body."
        404:
          description: "Project or requirement not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "Requirement management"
      summary: "Delete requirement"
      description: ""
      produces:
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "req_id"
        description: "requirement id"
        in: "path"
        type: "integer"
        required: true
      responses:
        200:
          description: "Operation successful."
        404:
          description: "Project or requirement not found."
        500:
          description: "Operation failed."
  /projects/{proj_id}/requirements/{req_id}/assign-tag:
    post:
      tags:
      - "Requirement management"
      summary: "Assign tag to requirement"
      description: ""
      produces:
      - "text/html"
      - "application/json"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "req_id"
        description: "requirement id"
        in: "path"
        type: "integer"
        required: true
      - name: "tag info"
        in: "body"
        description: "Tag ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            tag_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful. Returns updated requirement in JSON."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                $ref: "#/definitions/Requirement_full"
        400:
          description: "Invalid body."
        404:
          description: "Requirement or project not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "Requirement management"
      summary: "Remove assigned tag from requirement"
      description: ""
      produces:
      - "text/html"
      - "application/json"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "req_id"
        description: "requirement id"
        in: "path"
        type: "integer"
        required: true
      - name: "tag info"
        in: "body"
        description: "Tag ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            tag_id:
              type: "integer"
      responses:
        200:
          description: "Operation successful. Returns updated requirement in JSON."
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                $ref: "#/definitions/Requirement_full"
        400:
          description: "Invalid body."
        404:
          description: "Requirement or project not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
...
  /projects/{proj_id}/plans/{tp_id}/mods/{tm_id}/cases/{tc_id}/requirements/assign-requirement:
    post:
      tags:
      - "Requirement management"
      summary: "Assign requirement to test case."
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "tp_id"
        description: "test plan id"
        in: "path"
        type: "integer"
        required: true
      - name: "tm_id"
        description: "test module id"
        in: "path"
        type: "integer"
        required: true
      - name: "tc_id"
        description: "test case id"
        in: "path"
        type: "integer"
        required: true
      - name: "requirement info"
        in: "body"
        description: "Requirement ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            requirement_id:
              type: "integer"
      responses:
        200:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                type: "array"
                items:
                  $ref: "#/definitions/Requirement_full"
          description: "Operation successful. Returns test case current requirements in JSON response."
        400:
          description: "Invalid body."
        404:
          description: "Project, test plan, test module, test case or its requirements not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
    delete:
      tags:
      - "Requirement management"
      summary: "Deleted assigned requirement from test case."
      description: ""
      produces:
      - "application/json"
      - "text/html"
      parameters:
      - name: "proj_id"
        description: "project id"
        in: "path"
        type: "integer"
        required: true
      - name: "tp_id"
        description: "test plan id"
        in: "path"
        type: "integer"
        required: true
      - name: "tm_id"
        description: "test module id"
        in: "path"
        type: "integer"
        required: true
      - name: "tc_id"
        description: "test case id"
        in: "path"
        type: "integer"
        required: true
      - name: "requirement info"
        in: "body"
        description: "Requirement ID in JSON"
        required: true
        schema:
          type: "object"
          properties:
            requirement_id:
              type: "integer"
      responses:
        200:
          schema:
            type: "object"
            properties:
              message:
                type: "string"
              object:
                type: "array"
                items:
                  $ref: "#/definitions/Requirement_full"
          description: "Operation successful. Returns test case current requirements in JSON response."
        400:
          description: "Invalid body."
        404:
          description: "Project, test plan, test module, test case or its requirements not found."
        415:
          description: "Unsupported media type."
        500:
          description: "Operation failed."
definitions:
...
  TestStepDefinition:
    type: "object"
    required:
    - id
    - content
    - project
    - gherkin_type
    properties:
      id:
        type: "integer"
      content:
        type: "string"
      gherkin_type:
        type: "string"
        description: "expected name of step gherkin type"
      param_count:
        type: "integer"
      project:
        type: "object"
        properties:
          id:
            type: "integer"
          name:
            type: "string"
  TestStep:
    type: "object"
    required:
    - id
    - content
    - order
    - test_step_definition
    properties:
      id:
        type: "integer"
      content:
        type: "string"
      order:
        type: "integer"
      test_step_definition:
        $ref: "#/definitions/TestStepDefinition"
  TestCase:
    type: "object"
    required:
    - id
    - name
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      priority:
        enum: ["high", "medium", "low"]
        description: "Default priority types, connected to Test Case."
        type: "string"
      scenario_name:
        type: "string"
        description: "also 'scenario_no_spaces' is used for filenames based on test case"
      steps:
        type: "array"
        items:
          $ref: "#/definitions/TestStep"
...
  Tag:
    type: "object"
    required:
    - id
    - project_id
    properties:
      id:
        type: "integer"
      content:
        type: "string"
      category:
        type: "string"
      project_id:
        type: "string"
...
  Requirement_full:
    type: "object"
    required:
    - id
    - project_id
    properties:
      id:
        type: "integer"
      content:
        type: "string"
      project_id:
        type: "string"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
      project:
        type: "object"
        properties:
          id:
            type: "integer"
          name:
            type: "string"
      in_testcases:
        type: "array"
        items:
          $ref: "#/definitions/TestCase"